requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
babel==2.18.0
    # via mkdocs-material
backrefs==6.2
    # via mkdocs-material
certifi==2026.2.25
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via mkdocs
colorama==0.4.6
    # via mkdocs-material
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
fonttools==4.61.1
    # via matplotlib
ghp-import==2.1.0
    # via mkdocs
griffelib==2.0.0
    # via mkdocstrings-python
hjson==3.1.0
    # via
    #   mkdocs-macros-plugin
    #   super-collections
idna==3.11
    # via requests
importlib-metadata==8.7.1
    # via mike
importlib-resources==6.5.2
    # via mike
jinja2==3.1.6
    # via
    #   mike
    #   mkdocs
    #   mkdocs-macros-plugin
    #   mkdocs-material
    #   mkdocstrings
kiwisolver==1.4.9
    # via matplotlib
markdown==3.10.2
    # via
    #   -r requirements.in
    #   mkdocs
    #   mkdocs-autorefs
    #   mkdocs-material
    #   mkdocstrings
    #   pymdown-extensions
markupsafe==3.0.3
    # via
    #   jinja2
    #   mkdocs
    #   mkdocs-autorefs
    #   mkdocstrings
matplotlib==3.10.8
    # via
    #   ptitprince
    #   seaborn
mergedeep==1.3.4
    # via
    #   mkdocs
    #   mkdocs-get-deps
mike==2.1.3
    # via -r requirements.in
mkdocs==1.6.1
    # via
    #   -r requirements.in
    #   mike
    #   mkdocs-autorefs
    #   mkdocs-macros-plugin
    #   mkdocs-material
    #   mkdocstrings
mkdocs-autorefs==1.4.4
    # via
    #   mkdocstrings
    #   mkdocstrings-python
mkdocs-get-deps==0.2.0
    # via mkdocs
mkdocs-macros-plugin==1.5.0
    # via -r requirements.in
mkdocs-material==9.7.3
    # via -r requirements.in
mkdocs-material-extensions==1.3.1
    # via mkdocs-material
mkdocstrings[python]==1.0.3
    # via
    #   -r requirements.in
    #   mkdocstrings-python
mkdocstrings-python==2.0.3
    # via mkdocstrings
numpy==2.4.2
    # via
    #   contourpy
    #   matplotlib
    #   pandas
    #   ptitprince
    #   scipy
    #   seaborn
packaging==26.0
    # via
    #   matplotlib
    #   mkdocs
    #   mkdocs-macros-plugin
paginate==0.5.7
    # via mkdocs-material
pandas==3.0.1
    # via seaborn
pathspec==1.0.4
    # via
    #   mkdocs
    #   mkdocs-macros-plugin
pillow==12.1.1
    # via matplotlib
platformdirs==4.9.2
    # via mkdocs-get-deps
ptitprince==0.3.1
    # via -r requirements.in
pygments==2.19.2
    # via mkdocs-material
pymdown-extensions==10.21
    # via
    #   -r requirements.in
    #   mkdocs-material
    #   mkdocstrings
pyparsing==3.3.2
    # via
    #   matplotlib
    #   mike
python-dateutil==2.9.0.post0
    # via
    #   ghp-import
    #   matplotlib
    #   mkdocs-macros-plugin
    #   pandas
pyyaml==6.0.3
    # via
    #   mike
    #   mkdocs
    #   mkdocs-get-deps
    #   mkdocs-macros-plugin
    #   pymdown-extensions
    #   pyyaml-env-tag
pyyaml-env-tag==1.1
    # via
    #   mike
    #   mkdocs
requests==2.32.5
    # via
    #   mkdocs-macros-plugin
    #   mkdocs-material
scipy==1.17.1
    # via ptitprince
seaborn==0.13.2
    # via ptitprince
six==1.17.0
    # via python-dateutil
super-collections==0.6.2
    # via mkdocs-macros-plugin
termcolor==3.3.0
    # via mkdocs-macros-plugin
urllib3==2.6.3
    # via requests
verspec==0.1.0
    # via mike
watchdog==6.0.0
    # via mkdocs
zipp==3.23.0
    # via importlib-metadata
